Huiswerkopdracht: Endpoints optimaliseren. Muziekstuk: typo's in Javadoc

+++ src/main/java/nl/rollingsticks/rest/service/HuiswerkopdrachtEndpoint.java
@@ -15,6 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import nl.rollingsticks.domain.Huiswerkopdracht;
+import nl.rollingsticks.domain.model.MuziekstukModelBasic;
 import nl.rollingsticks.persistence.HuiswerkopdrachtService;
 
 @Path("huiswerkopdracht")
@@ -34,20 +35,40 @@ public class HuiswerkopdrachtEndpoint {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response postHuiswerkopdracht(Huiswerkopdracht huiswerkopdracht){
-		System.out.println("pre@POST: " + huiswerkopdracht.getId() + " - " + huiswerkopdracht.getNotitie());
+		System.out.println("Huiswerk - pre@POST: " + huiswerkopdracht.getId() + " - " + huiswerkopdracht.getNotitie());
 		Huiswerkopdracht result = huiswerkopdrachtService.save(huiswerkopdracht);
+		System.out.println("Huiswerk - @POST: " + huiswerkopdracht.getId() + " - " + huiswerkopdracht.getNotitie());
 		return Response.accepted(result).build();
 	}
 	
+	/**
+	 * Opvragen van de huiswerkopdracht.
+	 * Op basis van id worden de gegevens gefilterd via een JSON object teruggegeven.
+	 * @param 	id 	Id van de Huiswerkopdracht wordt uit het path gehaald.
+	 * @return 	Code 200 (OK)<br>
+	 * 		 	Code 204 (No Content)<br>
+	 * 			Opgevraagde Huiswerkopdracht wordt als JSON object teruggegeven.
+	 */	
 	@GET
 	@Produces(MediaType.APPLICATION_JSON)
 	@Path("{id}")
 	public Response getHuiswerkopdrachtById(@PathParam("id") Long id ) {
-		System.out.println("@GET: (" + id + ")");
-		Huiswerkopdracht result = this.huiswerkopdrachtService.findById(id);
-		return Response.ok(result).build();
+		System.out.println("Huiswerk - pre@GET: (" + id + ")");
+		Huiswerkopdracht huiswerkopdracht = this.huiswerkopdrachtService.findById(id);
+		if (huiswerkopdracht == null) {
+			System.out.println("Huiswerk - Id " + id + " bestaat niet.");
+			return Response.noContent().build();
+		} else {
+			System.out.println("Huiswerk - @GET: (" + id + ") " + huiswerkopdracht.getLesDatum());
+			return Response.ok(huiswerkopdracht).build();
+		}
 	}
 	
+	/**
+	 * Opvragen van alle Huiswerkopdrachten.
+	 * @return 	Code 200 (OK)<br>
+	 * 			Alle Huiswerkopdrachten worden als JSON objecten teruggegeven.
+	 */	
 	@GET
 	@Produces(MediaType.APPLICATION_JSON)
 	public Response listHuiswerkopdracht(){
@@ -56,12 +77,24 @@ public class HuiswerkopdrachtEndpoint {
 		return Response.ok(result).build();
 	}
 	
+	/**
+	 * Verwijderen van de opgegeven Huiswerkopdracht (id).
+	 * @param 	id 	Id van de te verwijderen Huiswerkopdracht wordt uit het path gehaald.
+	 * @return 	Code 202 (Accepted)<br>
+	 * 		 	Code 204 (No Content)
+	 */	
 	@DELETE
-	@Consumes(MediaType.APPLICATION_JSON)
 	@Path("{id}")
 	public Response deleteTekstById(@PathParam("id") Long id){
-		this.huiswerkopdrachtService.deleteById(id);
-		return Response.accepted().build();
+		System.out.println("Huiswerk - pre@DELETE: id provided: " + id);
+		Huiswerkopdracht huiswerkopdracht = this.huiswerkopdrachtService.findById(id);
+		if (huiswerkopdracht == null) {
+			System.out.println("Huiswerk - Id " + id + " bestaat niet.");
+			return Response.noContent().build();
+		} else {
+			this.huiswerkopdrachtService.deleteById(id);
+			return Response.accepted().build();
+		}
 	}
 
 	@PUT
